move icons folder from src to public
