+++ public/icons/AddMeasurementIcon.js
