+++ public/icons/AddPumpkinIcon.js
