+++ public/icons/SearchIcon.js
