+++ public/icons/SettingsIcon.js
