+++ src/components/GraphCard.js
@@ -2,7 +2,7 @@ import React, { useMemo, useRef } from 'react';
 import { Line } from 'react-chartjs-2';
 import 'chartjs-adapter-date-fns';
 import { Chart, registerables } from 'chart.js';
-import FullscreenIcon from './icons/FullscreenIcon';
+import FullscreenIcon from '../../public/icons/FullscreenIcon';
 
 Chart.register(...registerables);
