+++ src/components/ImageGallery.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import PlusIcon from './icons/PlusIcon';
+import PlusIcon from '../../public/icons/PlusIcon';
 import { useDropzone } from 'react-dropzone';
 
 const ImageGallery = ({ images, openModal, handleUpload }) => {
